A new option -u for interactively updating baselines.
While -U always copies the new output over the existing baselines, -u
will now ask for a failed test whether the users wants to update the
baseline or just continue. The check is done by btest-diff, and
there's a new helper btest-ask-update that does the actual interaction
with the user.

Two additional changes:

   - If a test return exit code 200, we now immediately abort btest's
     further execution.

   - Test processes receive a new environments variable TEST_NAME with
     the name of the current test.