Fix btest-bg-wait's kill trap and -k option.
The kill trap was trying to kill based on btest-bg-run's tag/directory
name instead of pids.

And when determining the exit code of btest-bg-wait itself, the -k option
was trying to factor in the exit codes of tasks that were killed due to
timing out.  Those cases are ignored now, but any task that exited on its
is still allowed to change the overall exit status.