Don't add btest's path to PATH anymore.
Because it impedes the ability to supply wrapper scripts around the other
btest utility scripts.  In a standard installation of btest, the user's
PATH should already be set correctly.  When running btest from the
distribution sources, the README indicates to add the distribution path
to PATH in the configuration file's environment section.