Separating semantics of groups and thread serialization into separate options.
Now -g still specifices TEST-GROUPs that are to be executed, but the
groups don't control which tests are serialized in in parallel
execution. For that, there's a new TEST-SERIALIZE command that takes a
tag and then makes sure all tests with the same tag are run in the
same thread.

TEST-GROUPS can also appear multiple times now to assign a test to a
set of groups. And -g takes a comma-separated list of groups names to
run more than one test group.