Adding a new "alternative" concept that combines filters and substitutions, and adds per-alternative environment variables.
Instead of defining filters and substitutions separately, one now
specifies an alternative configuration to run with "-A <name>" and
that then checks for both "[substitutions-<name>]" and
"[filter-<name>]" section. In addition, "[environment-<name>]" allows
to define alternative-specific environment variables.

The old filter/substitutions options -F and -s are gone. The sections
for substitutions are renamed to "[substitutions-<name>]" from
"[subst-<name>]".