Fix bug in btest-diff for certain canonifiers
Fixed a bug that would cause btest-diff to always return zero when
certain canonifiers are used.  Such canonifiers read a file when the
filename is supplied as a cmd-line argument instead of reading the file
from stdin (an example of this is the "sed" command).  The bug was
caused by btest-diff providing the canonifier with the wrong filename
on the cmd-line, and so the canonifier would ignore the baseline file.

As a result of this bug, some tests that use btest could appear to
succeed even if a command's output is totally different from its baseline.