Fix bug in btest-diff when a canonifier fails
Fixed a bug in btest-diff that would occur when a canonifier fails and
then btest-diff would just return zero (indicating success).  This
typically occurs when the path or filename of the canonifier is incorrect.
This is now fixed by checking the exit status of the canonifier command
and then showing a detailed error message upon failure.

As a result of this bug, some tests that use btest-diff could appear
to succeed even if a command's output is totally different from its
baseline.