Show fewer lines of file when btest-diff fails
Changed MAX_LINES in btest-diff to a smaller value to reduce
the number of lines of the file shown when a test fails.  The
previous value was so large that files were almost never
truncated in the output.

Added ability for a user to override this value by setting the
env. var. TEST_DIFF_FILE_MAX_LINES.

Finally, when no baseline exists, changed btest-diff to always just
show the entire file (because in this case there is no diff output,
and seeing the last part of the file can be useful).