New utility btest-progress to display progress messages while a test is executing.
This helps to track where long-running tests current are at.

Example:

    % cat example.test
    # @TEST-EXEC: bash %INPUT

    btest-progress Stage 1
    sleep 1
    btest-progress Stage 2
    sleep 1
    btest-progress Stage 3
    sleep 1

Output:

    % btest example.test
    example ...
      - Stage 1
      - Stage 2
      - Stage 3
    example ... ok
    all 1 tests successful

These messages appear in real-time while the rest is still running.
When stdout is a tty, the progress messages are incorporated into the
colored one-line status message.

By default, btest-progress also print the message to the tests
standard output. That can be suppressed by giving it an option -q.