Experimental automatic generation of test reference documentation.
The commit adds a new command-line option "-R <format>" that prints
out a list of all tests in either Markdown (format 'md') or
reStructuredText (format 'rst'). The list includes a documentation
string for each tests that gets defined through a new "@TEST-DOC:
<docstring>" directive.

Example output for the btest test-suite (which doesn't define many
@TEST-DOC yet):

    # btest -R rst
    tests
    -----

        ``tests.alternatives-environment``:
            No documentation.

        ``tests.alternatives-filter``:
            No documentation.

        [...]

        ``tests.progress``:
            Ensures that "btest-diff progress" functions correctly.

        [...]

    tests.sphinx
    ------------

        ``tests.sphinx.rst-cmd``:
            No documentation.

        ``tests.sphinx.run-sphinx``:
            No documentation.

This is experimental in terms of both current format and content.