Prevent socket path length from being too long
Since TMPDIR is assigned a value in many btest.cfg files (it typically
has the same path as TmpDir), the path length can be too long.  Solved
the problem by checking the pathname length.  If it is too long, then
btest will try to use the relative path of TmpDir (which will be ".tmp"
unless someone changed the value of TmpDir in btest.cfg).  If that's
still too long, then btest just uses the current directory.