Fix TTY output test to work on Linux and BSD
This specifically tests whether script supports the -c option (which
only exists on Linux) and uses this to distinguish BSD- and Linux-
specific implementations of the script invocation. The BSD
requirements test uses negation of the Linux one because running
script without -c on Linux causes it to enter a transcript session,
hanging btest.