Improve testing of setting a non-default base directory
Added more test cases for testing of setting a non-default base
directory (both relative and absolute paths are now being tested).

Also fixed the alternatives-testbase.test, which was broken for
two reasons:  files were being created (such as .btest.failed.dat)
in "testing/Files/local_alternative", and it failed to detect a bug in
btest.  The base dir name used in the test was
"../../Files/local_alternative", which btest was erroneously prepending
twice to the test name (the test couldn't detect the btest bug due to
the ".." components in the path).