Fix a bug when setting base dir using a relative path
When specifying a non-default base directory (using BTEST_CFG,
BTEST_TEST_BASE, or "btest -c"), btest was not converting the
path to an absolute path.  This along with a redundant line of code
that was joining the base dir and test filename resulted in the
base dir being prepended twice to the test filename.

Addresses BIT-1892.