Improve a unicode decode error message
When using python3, btest terminates with a python stack trace if a
unicode decode error occurs when reading the test files, but the name
of the problem test file is not shown.

Improved the error message by catching the exception and including the
filename of the problem test file in the error message.