Provide the name of the next upcoming test via TEST_NAME_NEXT env var
This allows tests to adjust their behavior depending on upcoming
tests. For example, tests that would individually engage in expensive
set-up/tear-down may skip some of that work if they know that the next
test will just re-establish conditions that the current test would
tear down.

For the final test, TEST_NAME_NEXT remains unset.

Includes a bit of refactoring to separate the lock-guarded
identification of the next suitable test, which modifies
TestManager._tests, from the derivation of the actual test(s) to run,
and a minor bugfix in TestManager.threadRun(), where it looks like we
invoked saveTiming() repeatedly on the same tests (due to the use of
all_tests).