Fix missing return value in get_env_from_alternative
If a BTEST_TEST_BASE environment variable is set, it should override
such a value if defined by alternatives. get_env_from_alternative()
forgot to return the default in that case.

This adds a test case to trigger that scenario, by expanding the
alternatives-testbase test: select an alternative with a
BTEST_TEST_BASE, but override its value via an env var.