Make `btest-progress` more robust under concurrency.
`btest-progress` generates files capturing progress. These were
previoulsy tagged with the number of seconds since the epoch, e.g., to
allow sorting outputs by time.

If test processes start at the same time this can lead to clashes on the
name, and e.g., different test processes writing to the same progress
file. In fact our test `tests.progress` already inserted artifical
sleeps to make this less likely to occur in the that particular test
(without addressing the underlying issue).

With this patch we now use nanosecond resolution for the names of the
temporary files. We additionally make sure we use unique files for each
invocation of `btest-progress` via `mktemp`. We also remove the
workaround in `tests.progress` and make use of parallelism in the
test Makefile.