Add support for Teardown, a run-always Finalizer
We're using btest to run system tests in a rather stateful environment
and were looking for a hook / entry point to run after every test for
cleanup and state checking.

Finalizer almost met the bill, except for not run for failing tests.

If running the executable given by Teardown fails, the whole test is
considered failed. If Teardown succeeds, but the test failed, the
test outcome is still failure.