Avoid assertion failure if same test required multiple times.
If one requests the same test multiple times we previously would
fail with an assertion error inside `mergeTestParts` which validates
that different tests cannot have the same name. This was due to
`findTests` emitting multiple instances of the same test, one for each
instance requested on the command line.

With this patch `findTests` now interally uses a `set` to store tests
which causes them to be deduplicated.