Make test `duplication-selection` independent of actual scheduling.
This test validates that only a single test is run exactly once, even if
requested multiple times. To validate that, we previously diffed the
full BTest log, but since this includes the worker ID it might produce
diffing failures even though the test should have passed.

With this patch we remove the output diffing and instead just validate
that the test is reported to succeed once.