Do not ignore case of options.
Options in our ini file are usually interpreted as environment variables
which are case-sensitive on many platforms, e.g., `http_proxy` and
`HTTP_PROXY` are different environment variables. Our config parser
reads in environment variables as configuration value defaults, but the
`ConfigParser` used under the covers prefers to treat them as
case-insensitve in its default ini file dialect.

With this patch we now override this behavior of our config parser and
treat config keys as case-sensitive. This yields the expected behavior
for environment variables, but also is a API-breaking change for our ini
files where we preserve case as well.

Closes #75.