Prevent console.test from garbling output
When running btest's own tests parallelized and without a TTY (e.g. "btest -j |
cat"), the script command invoked in the console.test breaks newlines: they no
longer carriage-return, leading to a staircase. I'm seeing this on Linux and
FreeBSD. I tried a bunch of approaches and in the end convincing script that it
doesn't actually run in a TTY looks like the easiest fix.

ps: I am fairly proud of this