Prevent output garbling when running with "-A -j"
The Console output handler (which kicks in when using -A) wasn't prepared for
parallelism. This commit switches it to using the base class's buffering logic,
so each test's output gets buffered up and written out as one when done,
but preserves outputting to stdout (thx Arne).

To keep CompactConsole working, this requires moving some of the method
implementations from the original Console implementation down into it.

This also removes the `show_all` members in the two handler classes since
they're fixed over the lifetime of their instances and weren't used much.