monitor_cb: Start thread with daemon=True
When an unexpected exception happened in a child process, nothing would
stop the monitor_cb thread and in turn the child process would never
finish. Use daemon=True on the monitor_cb thread to prevent this.

There should never be unhandled exceptions in child processes, because
it may then not close all resources properly etc, but a non-terminating
btest is worse, specifically when btest is run in CI unattended.