stderr/stdout: Do not assume well-formed UTF-8 output
The previous code opened stdout/stderr providing an explicit encoding="utf-8"
argument. This has no effect when the file handles are passed as stderr/stdout
to a subprocess: Only the underlying fds are passed and there's no way to
control the encoding (if any) used by the subprocess for the bytes it
produces on these streams.

So, switch things around: Remove the misleading encoding arguments for the
stdout, stderr files, open them in binary mode and be careful when
opening .stderr for displaying purposes. Do try UTF-8, but also set error
strategy to "backslashreplace" to avoid errors.